update to fileMatchHash
