+++ lib/workers/repository/extract/index.spec.ts
@@ -1,7 +1,6 @@
 import { defaultConfig, git, mocked } from '../../../../test/util';
 import type { RenovateConfig } from '../../../config/types';
 import { logger } from '../../../logger';
-import { getExtractList } from '../process/extract-update';
 import * as _managerFiles from './manager-files';
 import { extractAllDependencies } from '.';
 
@@ -23,32 +22,32 @@ describe('workers/repository/extract/index', () => {
 
     it('runs', async () => {
       managerFiles.getManagerPackageFiles.mockResolvedValue([{} as never]);
-      const [extractList] = await getExtractList(config);
-      const res = await extractAllDependencies(config, extractList);
+
+      const res = await extractAllDependencies(config);
       expect(Object.keys(res)).toContain('ansible');
     });
 
     it('skips non-enabled managers', async () => {
       config.enabledManagers = ['npm'];
       managerFiles.getManagerPackageFiles.mockResolvedValue([{} as never]);
-      const [extractList] = await getExtractList(config);
-      const res = await extractAllDependencies(config, extractList);
+
+      const res = await extractAllDependencies(config);
       expect(res).toEqual({ npm: [{}] });
     });
 
     it('warns if no packages found for a enabled manager', async () => {
       config.enabledManagers = ['npm'];
       managerFiles.getManagerPackageFiles.mockResolvedValue([]);
-      const [extractList] = await getExtractList(config);
-      expect(await extractAllDependencies(config, extractList)).toEqual({});
+
+      expect(await extractAllDependencies(config)).toEqual({});
       expect(logger.debug).toHaveBeenCalled();
     });
 
     it('warns if packageFiles is null', async () => {
       config.enabledManagers = ['npm'];
       managerFiles.getManagerPackageFiles.mockResolvedValue(null);
-      const [extractList] = await getExtractList(config);
-      expect(await extractAllDependencies(config, extractList)).toEqual({});
+
+      expect(await extractAllDependencies(config)).toEqual({});
     });
 
     it('adds skipReason to internal deps when updateInternalDeps is false/undefined', async () => {
@@ -58,8 +57,8 @@ describe('workers/repository/extract/index', () => {
           deps: [{ depName: 'a', isInternal: true }, { depName: 'b' }],
         },
       ]);
-      const [extractList] = await getExtractList(config);
-      expect(await extractAllDependencies(config, extractList)).toEqual({
+
+      expect(await extractAllDependencies(config)).toEqual({
         npm: [
           {
             deps: [
@@ -79,8 +78,8 @@ describe('workers/repository/extract/index', () => {
     it('checks custom managers', async () => {
       managerFiles.getManagerPackageFiles.mockResolvedValue([{} as never]);
       config.regexManagers = [{ fileMatch: ['README'], matchStrings: [''] }];
-      const [extractList] = await getExtractList(config);
-      const res = await extractAllDependencies(config, extractList);
+
+      const res = await extractAllDependencies(config);
       expect(Object.keys(res)).toContain('regex');
     });
   });
